Only impact openapi and watch

+++ src/Shared/Process/ProcessExtensions.cs
@@ -30,7 +30,86 @@ namespace Microsoft.Extensions.Internal
             }
             else
             {
+                var children = new HashSet<int>();
+                GetAllChildIdsUnix(pid, children, timeout);
+                foreach (var childId in children)
+                {
+                    KillProcessUnix(childId, timeout);
+                }
+                KillProcessUnix(pid, timeout);
+            }
+        }
+
+        public static void KillTreeNoPgrep(this Process process) => process.KillTreeNoPgrep(_defaultTimeout);
+
+        public static void KillTreeNoPgrep(this Process process, TimeSpan timeout)
+        {
+            var pid = process.Id;
+            if (_isWindows)
+            {
+                RunProcessAndWaitForExit(
+                    "taskkill",
+                    $"/T /F /PID {pid}",
+                    timeout,
+                    out var _);
+            }
+            else
+            {
                 process.Kill(entireProcessTree: true);
+            }
+        }
+
+        private static void GetAllChildIdsUnix(int parentId, ISet<int> children, TimeSpan timeout)
+        {
+            try
+            {
+                RunProcessAndWaitForExit(
+                    "pgrep",
+                    $"-P {parentId}",
+                    timeout,
+                    out var stdout);
+
+                if (!string.IsNullOrEmpty(stdout))
+                {
+                    using (var reader = new StringReader(stdout))
+                    {
+                        while (true)
+                        {
+                            var text = reader.ReadLine();
+                            if (text == null)
+                            {
+                                return;
+                            }
+
+                            if (int.TryParse(text, out var id))
+                            {
+                                children.Add(id);
+                                // Recursively get the children
+                                GetAllChildIdsUnix(id, children, timeout);
+                            }
+                        }
+                    }
+                }
+            }
+            catch (Win32Exception ex) when (ex.Message.Contains("No such file or directory"))
+            {
+                // This probably means that pgrep isn't installed. Nothing to be done?
+            }
+        }
+
+        private static void KillProcessUnix(int processId, TimeSpan timeout)
+        {
+            try
+            {
+                RunProcessAndWaitForExit(
+                    "kill",
+                    $"-TERM {processId}",
+                    timeout,
+                    out var stdout);
+            }
+            catch (Win32Exception ex) when (ex.Message.Contains("No such file or directory"))
+            {
+                // This probably means that the process is already dead
             }
         }
